src/ru/mail/avdienkoartyom/storage/AbstractArrayStorage.java
package ru.mail.avdienkoartyom.storage;

import ru.mail.avdienkoartyom.exception.ExistStorageException;
import ru.mail.avdienkoartyom.exception.NoExistStorageException;
import ru.mail.avdienkoartyom.exception.StorageException;
import ru.mail.avdienkoartyom.model.Resume;

import java.util.Arrays;

public abstract class AbstractArrayStorage implements Storage {
    protected static final int STORAGE_LIMIT = 10_000;
    protected final Resume[] storage = new Resume[STORAGE_LIMIT];
    protected int size = 0;

    protected abstract int getIndex(String uuid);

    protected abstract void deleteElement(int index);

    protected abstract void saveElement(Resume resume, int index);

    public void save(Resume resume) {
        int index = getIndex(resume.getUuid());
        if (size == STORAGE_LIMIT) {
            throw new StorageException("Storage Overflow", resume.getUuid());
        } else if (index < 0) {
            if (size == 0) {
                storage[size] = resume;
            } else {
                saveElement(resume, index);
            }
            size++;
        } else {
            throw new ExistStorageException(resume.getUuid());
        }

    }

    public void delete(String uuid) {
        int index = getIndex(uuid);
        if (index > -1) {
            deleteElement(index);
            storage[size - 1] = null;
            size--;
        } else {
            throw new NoExistStorageException(uuid);
        }
    }

    public Resume get(String uuid) {
        int index = getIndex(uuid);
        if (index > -1) {
            return storage[index];
        } else {
            throw new NoExistStorageException(uuid);
        }
    }

    public void update(Resume resume) {
        int index = getIndex(resume.getUuid());
        if (index > -1) {
            storage[index] = resume;
        } else {
            throw new NoExistStorageException(resume.getUuid());
        }
    }

    public Resume[] getAll() {
        return Arrays.copyOfRange(storage, 0, size);
    }

    public int size() {
        return size;
    }

    public void clear() {
        Arrays.fill(storage, 0, size, null);
        size = 0;
    }
}
